Began implementing global, transparent device scheduling. Allows all Runtimes to acquire wakelocks and schedule device wakeups across all platforms.

+++ dispatch/core/RemoteFunction/RemoteFunction.hh
@@ -94,7 +94,7 @@ class RemoteFunction
                 return nullptr;
             }
 
-            int parametersLength = sizeof...(params);
+            size_t parametersLength = sizeof...(params);
             if(parametersLength != mutatorHelpers.size())
             {
                 Logger::logError("Failed to execute RemoteFunction (RPC stub) \"%s\". "
@@ -213,7 +213,7 @@ class RemoteFunction
 
             std::string parameterNames = this->mutatorHelpers.size() > 0 ? this->mutatorHelpers[0]->getTypeName() : "";
 
-            for(int i = 1; i < this->mutatorHelpers.size(); i++)
+            for(size_t i = 1; i < this->mutatorHelpers.size(); i++)
             {
                 parameterNames += ", " + this->mutatorHelpers[i]->getTypeName();
             }
