+++ dispatch/core/RemoteFunction/RemoteFunctionRunnable.hh
@@ -163,7 +163,7 @@ class RemoteFunctionRunnable : public AbstractRemoteFunctionRunnable
 
             std::string parameterNames = this->mutatorHelpers.size() > 0 ? this->mutatorHelpers[0]->getTypeName() : "";
 
-            for(int i = 1; i < this->mutatorHelpers.size(); i++)
+            for(size_t i = 1; i < this->mutatorHelpers.size(); i++)
             {
                 parameterNames += ", " + this->mutatorHelpers[i]->getTypeName();
             }
